requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
litellm
python-dotenv